+++ packages/robo/src/core/env.ts
@@ -11,6 +11,7 @@ export const env = {
 	nodeEnv: process.env.NODE_ENV,
 	roboplay: {
 		api: process.env.ROBOPLAY_API ?? 'https://api.roboplay.dev',
+		debug: process.env.ROBOPLAY_DEBUG === 'true',
 		env: process.env.ROBOPLAY_ENV,
 		frontend: process.env.ROBOPLAY_FRONTEND ?? 'https://roboplay.dev',
 	}